a handful of changes
Main issue solved is placebid function in server it now handles all eventualities if the auction is over or still ongoing.
changes int, int32 and int64 to just be int64 because using all three makes the code harder to work with and read and with the scope as small as it is the downsides of 64 are minimal.

movse mutex lock to ITU_databseserver